services/detection-engine/requirements.txt
# SENTINAL Detection Engine — Python dependencies
# Updated: integrated sentinel-ml ML model + Gemini explainer

# ─────────────────────────────────────────────────────────────────
# Core FastAPI stack (unchanged)
# ─────────────────────────────────────────────────────────────────
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.1
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.24
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
orjson==3.11.7
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.2
python-multipart==0.0.22
PyYAML==6.0.3
rich==14.3.3
rich-toolkit==0.19.7
shellingham==1.5.4
starlette==0.37.2
typer==0.24.1
typing-inspection==0.4.2
typing_extensions==4.15.0
ujson==5.12.0
uvicorn==0.29.0
uvloop==0.22.1
watchfiles==1.1.1
websockets==16.0

# ─────────────────────────────────────────────────────────────────
# NEW — sentinel-ml integration (ML model + Gemini explainer)
# Added: 2026-03-31 — feature/integrate-sentinel-ml
# ─────────────────────────────────────────────────────────────────

# ML model loading — required to deserialise sentinel_v5.pkl
joblib>=1.3.0

# XGBoost — sentinel_v5.pkl was trained with this (must be compatible version)
# Pin to >=2.0.0 which introduced the sklearn-compatible API used in training
xgboost>=2.0.0

# scikit-learn — required by XGBoost's sklearn wrapper + predict_proba()
# joblib also pulls this in, but explicit pin ensures version compat
scikit-learn>=1.3.0

# Gemini LLM explainer — only used when GEMINI_API_KEY is set in .env
# If key is absent, this package is installed but never imported at runtime
google-generativeai>=0.8.0
